+++ examples/hello-world/main.ts
@@ -1,5 +1,5 @@
 import { Client } from "@worlds/client";
-import { ComunicaSparqlEngine } from "@worlds/client/adapters/comunica";
+import { createComunicaSparqlEngineFactory } from "@worlds/client/adapters/comunica";
 import { createRdfjsClientOptions } from "@worlds/client/adapters/rdfjs";
 import { QueryEngine } from "@comunica/query-sparql-rdfjs-lite";
 
@@ -7,8 +7,7 @@ if (import.meta.main) {
   const queryEngine = new QueryEngine();
   const client = new Client(
     createRdfjsClientOptions({
-      createSparqlEngine: ({ store }) =>
-        new ComunicaSparqlEngine({ queryEngine, store }),
+      createSparqlEngine: createComunicaSparqlEngineFactory({ queryEngine }),
     }),
   );