+++ src/client/adapters/comunica/mod.ts
@@ -1,1 +1,2 @@
 export * from "./comunica-sparql-engine.ts";
+export * from "./create-comunica-sparql-engine-factory.ts";